Refactor proxy code
