+++ internal/proxy/utils.go
@@ -1,0 +1,28 @@
+package proxy
+
+import (
+	"net/url"
+	"strings"
+)
+
+// joinURLPath joins two URL paths together, handling edge cases with slashes
+func joinURLPath(a, b *url.URL) (path, rawpath string) {
+	apath := a.EscapedPath()
+	if apath == "" {
+		apath = "/"
+	}
+	bpath := b.EscapedPath()
+	if bpath == "" {
+		bpath = "/"
+	}
+	if strings.HasSuffix(apath, "/") && strings.HasPrefix(bpath, "/") {
+		apath = apath + bpath[1:]
+	} else {
+		apath = apath + bpath
+	}
+	unescaped, err := url.PathUnescape(apath)
+	if err != nil {
+		return apath, apath
+	}
+	return unescaped, apath
+}
